WIP: Peer management from the application settings

+++ app/src/main/java/io/sikorka/android/ui/wizard/slides/network_selection/NetworkSelectionContract.kt
@@ -10,6 +10,6 @@ interface NetworkSelectionView : BaseView {
 
 
 interface NetworkSelectionPresenter : Presenter<NetworkSelectionView> {
-  fun selectNetwork(@Network.Selection network: Long)
+  fun selectNetwork(@Network.Selection network: Int)
   fun updateSelected()
 }
